Desktop/UniFinder/requirements.txt
Flask
google-genai
waitress
